Revert "Handle GET /source/serialized_suite/{serializedSuiteId} (#260)"
This reverts commit a5b915771911360cc8d4a04449cddff0a2472d60.
