--- tests/Functional/Client/SerializedSuiteClient/GetTest.php
@@ -1,71 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace SmartAssert\ApiClient\Tests\Functional\Client\SerializedSuiteClient;
-
-use GuzzleHttp\Psr7\Response;
-use Psr\Http\Message\ResponseInterface;
-use SmartAssert\ApiClient\Tests\Functional\Client\ClientActionThrowsIncompleteDataExceptionTestTrait;
-use SmartAssert\ApiClient\Tests\Functional\Client\ExpectedRequestProperties;
-use SmartAssert\ApiClient\Tests\Functional\Client\RequestAuthenticationTestTrait;
-use SmartAssert\ApiClient\Tests\Functional\Client\RequestPropertiesTestTrait;
-use SmartAssert\ApiClient\Tests\Functional\DataProvider\InvalidJsonResponseExceptionDataProviderTrait;
-use SmartAssert\ApiClient\Tests\Functional\DataProvider\NetworkErrorExceptionDataProviderTrait;
-
-class GetTest extends AbstractSerializedSuiteClientTestCase
-{
-    use ClientActionThrowsIncompleteDataExceptionTestTrait;
-    use InvalidJsonResponseExceptionDataProviderTrait;
-    use NetworkErrorExceptionDataProviderTrait;
-    use RequestPropertiesTestTrait;
-    use RequestAuthenticationTestTrait;
-
-    public static function clientActionThrowsExceptionDataProvider(): array
-    {
-        return array_merge(
-            self::networkErrorExceptionDataProvider(),
-            self::invalidJsonResponseExceptionDataProvider(),
-        );
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public static function incompleteResponseDataExceptionDataProvider(): array
-    {
-        return [
-            'id missing' => [
-                'payload' => ['suite_id' => self::SUITE_ID, self::STATE => 'requested', 'parameters' => []],
-                'expectedRequestName' => 'get_serialized-suite-get',
-                'expectedMissingKey' => 'id',
-            ],
-        ];
-    }
-
-    protected function createClientActionCallable(): callable
-    {
-        return function () {
-            $this->client->get(self::API_KEY, self::ID);
-        };
-    }
-
-    protected function getResponseFixture(): ResponseInterface
-    {
-        return new Response(
-            200,
-            ['content-type' => 'application/json'],
-            (string) json_encode([
-                'id' => self::ID,
-                'suite_id' => self::SUITE_ID,
-                'state' => self::STATE,
-                'tests' => [],
-            ])
-        );
-    }
-
-    protected function getExpectedRequestProperties(): ExpectedRequestProperties
-    {
-        return new ExpectedRequestProperties('GET', '/source/serialized_suite/' . self::ID);
-    }
-}
